Rename Upcoming Interview Validation Rules to match form of Activity Form rules
